Add permission callback to rest endpoint
fixes the `Function register_rest_route was called incorrectly. The REST API route definition for xeno-canto-oembed/v1/embed is missing the required permission_callback argument. For REST API routes that are intended to be public, use __return_true as the permission callback. (This message was added in version 5.5.0.)` error message.